refactor: use psr4 autoloader
