+++ src/Command/GarbageCollectCommand.php
