+++ src/Docker/Configuration/Adapter.php
