+++ src/Docker/Configuration/Container/Adapter.php
