+++ src/Docker/Configuration/Image.php
