+++ src/Docker/Configuration/State/Adapter.php
