+++ src/Docker/Configuration/Usage/Adapter.php
