+++ src/Docker/Image.php
