+++ src/Docker/Image/DockerHub/PrivateRepository.php
@@ -1,44 +1,63 @@
 <?php
 
-namespace Keboola\DockerBundle\Docker\Image;
+namespace Keboola\DockerBundle\Docker\Image\DockerHub;
 
-use Aws\Ecr\EcrClient;
-use Aws\Ecr\Exception\EcrException;
-use Aws\Exception\CredentialsException;
 use Keboola\DockerBundle\Docker\Component;
 use Keboola\DockerBundle\Docker\Image;
+use Keboola\DockerBundle\Exception\ApplicationException;
 use Keboola\DockerBundle\Exception\LoginFailedException;
 use Keboola\ObjectEncryptor\ObjectEncryptor;
-use Keboola\Syrup\Exception\ApplicationException;
 use Psr\Log\LoggerInterface;
 use Retry\BackOff\ExponentialBackOffPolicy;
 use Retry\Policy\SimpleRetryPolicy;
 use Retry\RetryProxy;
 use Symfony\Component\Process\Process;
 
-class AWSElasticContainerRegistry extends Image
+class PrivateRepository extends Image\DockerHub
 {
-    protected $awsRegion = 'us-east-1';
+    protected $loginUsername;
+    protected $loginPassword;
+    protected $loginServer;
 
     public function __construct(ObjectEncryptor $encryptor, Component $component, LoggerInterface $logger)
     {
         parent::__construct($encryptor, $component, $logger);
-        if (!empty($component->getImageDefinition()["repository"]["region"])) {
-            $this->awsRegion = $component->getImageDefinition()["repository"]["region"];
+        $config = $component->getImageDefinition();
+        if (isset($config["repository"])) {
+            if (isset($config["repository"]["username"])) {
+                $this->loginUsername = $config["repository"]["username"];
+            }
+            if (isset($config["repository"]["#password"])) {
+                $this->loginPassword = $this->getEncryptor()->decrypt($config["repository"]["#password"]);
+            }
+            if (isset($config["repository"]["server"])) {
+                $this->loginServer = $config["repository"]["server"];
+            }
         }
     }
 
     /**
      * @return string
      */
-    public function getAwsRegion()
+    public function getLoginUsername()
     {
-        return $this->awsRegion;
+        return $this->loginUsername;
     }
 
-    public function getAwsAccountId()
+    /**
+     * @return string
+     */
+    public function getLoginPassword()
     {
-        return substr($this->getImageId(), 0, strpos($this->getImageId(), '.'));
+        return $this->loginPassword;
+    }
+
+    /**
+     * @return string
+     */
+    public function getLoginServer()
+    {
+        return $this->loginServer;
     }
 
     /**
@@ -46,23 +65,17 @@ class AWSElasticContainerRegistry extends Image
      */
     public function getLoginParams()
     {
-        $ecrClient = new EcrClient(array(
-            'region' => $this->getAwsRegion(),
-            'version' => '2015-09-21'
-        ));
-        try {
-            $authorization = $ecrClient->getAuthorizationToken(["registryIds" => [$this->getAwsAccountId()]]);
-        } catch (CredentialsException $e) {
-            throw new LoginFailedException($e->getMessage());
-        } catch (EcrException $e) {
-            throw new LoginFailedException($e->getMessage());
+        // Login
+        $loginParams = [];
+        if ($this->getLoginUsername()) {
+            $loginParams[] = "--username=" . escapeshellarg($this->getLoginUsername());
         }
-        // decode token and extract user
-        list($user, $token) = explode(":", base64_decode($authorization->get('authorizationData')[0]['authorizationToken']));
-
-        $loginParams[] = "--username=" . escapeshellarg($user);
-        $loginParams[] = "--password=" . escapeshellarg($token);
-        $loginParams[] = escapeshellarg($authorization->get('authorizationData')[0]['proxyEndpoint']);
+        if ($this->getLoginPassword()) {
+            $loginParams[] = "--password=" . escapeshellarg($this->getLoginPassword());
+        }
+        if ($this->getLoginServer()) {
+            $loginParams[] = escapeshellarg($this->getLoginServer());
+        }
         return join(" ", $loginParams);
     }
 
@@ -72,7 +85,9 @@ class AWSElasticContainerRegistry extends Image
     public function getLogoutParams()
     {
         $logoutParams = [];
-        $logoutParams[] = escapeshellarg($this->getImageId());
+        if ($this->getLoginServer()) {
+            $logoutParams[] = escapeshellarg($this->getLoginServer());
+        }
         return join(" ", $logoutParams);
     }
 
@@ -84,7 +99,6 @@ class AWSElasticContainerRegistry extends Image
         $retryPolicy = new SimpleRetryPolicy(3);
         $backOffPolicy = new ExponentialBackOffPolicy(10000);
         $proxy = new RetryProxy($retryPolicy, $backOffPolicy);
-
         $command = "sudo docker run --rm -v /var/run/docker.sock:/var/run/docker.sock " .
             "docker:1.11 sh -c " .
             escapeshellarg(
@@ -100,8 +114,11 @@ class AWSElasticContainerRegistry extends Image
             });
             $this->logImageHash();
         } catch (\Exception $e) {
-            if (strpos($process->getOutput(), "403 Forbidden") !== false) {
+            if (strpos($process->getOutput(), "Login with your Docker ID to push and pull images from Docker Hub.") !== false) {
                 throw new LoginFailedException($process->getOutput());
+            }
+            if (strpos($process->getErrorOutput(), "unauthorized: incorrect username or password") !== false) {
+                throw new LoginFailedException($process->getErrorOutput());
             }
             throw new ApplicationException("Cannot pull image '{$this->getFullImageId()}': ({$process->getExitCode()}) {$process->getErrorOutput()} {$process->getOutput()}", $e);
         }
