+++ src/Docker/Image/QuayIO.php
