+++ src/Docker/Image/QuayIO/PrivateRepository.php
