+++ src/Docker/ImageFactory.php
