+++ src/Docker/OutputFilter/NullFilter.php
