+++ src/Docker/OutputFilter/OutputFilter.php
