+++ src/Docker/OutputFilter/OutputFilterInterface.php
