+++ src/Exception/BuildException.php
@@ -2,11 +2,6 @@
 
 namespace Keboola\DockerBundle\Exception;
 
-use Keboola\Syrup\Exception\ApplicationException;
-
-/**
- * Class BuildException
- */
 class BuildException extends ApplicationException
 {
 }
