+++ src/Exception/BuildParameterException.php
@@ -2,11 +2,6 @@
 
 namespace Keboola\DockerBundle\Exception;
 
-use Keboola\Syrup\Exception\UserException;
-
-/**
- * Class BuildException
- */
 class BuildParameterException extends UserException
 {
 }
