+++ src/Exception/OutOfMemoryException.php
@@ -2,11 +2,6 @@
 
 namespace Keboola\DockerBundle\Exception;
 
-use Keboola\Syrup\Exception\UserException;
-
-/**
- * Class LoginFailedException
- */
 class OutOfMemoryException extends UserException
 {
 }
