+++ src/Monolog/Processor/DockerContainerProcessor.php
