+++ src/Monolog/Processor/DockerProcessor.php
